% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TOLLoutput.R
\docType{methods}
\name{plot,TROLLoutput,ANY-method}
\alias{plot,TROLLoutput,ANY-method}
\title{Function to plot TROLL outputs.}
\usage{
\S4method{plot}{TROLLoutput,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{TROLLoutput}

\item{y}{char. Output to plot, see details.}
}
\value{
Plot the output
}
\description{
Function to plot TROLL outputs.
}
\details{
Available plots
\describe{
\item{defaults}{Not implemented}
\item{abund}{To precise}
\item{agb}{To precise}
\item{ba}{To precise}
\item{death}{To precise}
\item{final pattern}{To precise}
\item{flux}{To precise}
\item{gpp}{To precise}
\item{LAI}{To precise}
\item{litterfall}{To precise}
\item{npp}{To precise}
\item{R}{To precise}
\item{Ra}{To precise}
\item{relabund}{To precise}
\item{vertd}{To precise}
\item{wd}{To precise}
}
}
\examples{

}
