% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TROLLoutput.R
\docType{methods}
\name{print,TROLLoutput-method}
\alias{print,TROLLoutput-method}
\title{Function to print TROLL outputs.}
\usage{
\S4method{print}{TROLLoutput}(x, ...)
}
\arguments{
\item{x}{TROLLoutput}
}
\value{
Print in console
}
\description{
Function to print TROLL outputs.
}
\examples{

}
