% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run TROLL}
\usage{
run(name, input = "./src/input.txt", path = "./src/OUTPUT",
  overwrite = FALSE)
}
\arguments{
\item{name}{char. name of the model}

\item{input}{char. input file path}

\item{path}{char. apth to save the ouptuts}

\item{overwrite}{logical. allow to overwrite existing outputs files}
}
\value{
model output files in the path folder
}
\description{
Function to run TROLL code
}
\examples{

}
